[Move] ColletionView 위치 이동 #99
