+++ PPPCLUB-iOS/Presentation/Home/View/HomeView.swift
@@ -18,16 +18,15 @@ final class HomeView: UIView {
     // MARK: - UI Components
     
     public lazy var homeNavigationView = HomeNavigationView()
-    public lazy var homeWeeklyView = HomeWeeklyView()
     public lazy var homeAllView = HomeAllView()
-    var weeklyCollectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout())
+    
+    public lazy var homeWeeklyView = HomeWeeklyView()
+    
     
     // MARK: - Life Cycle
     
     override init(frame: CGRect) {
         super.init(frame: frame)
-        
-        register()
         
         style()
         hierarchy()
@@ -35,6 +34,15 @@ final class HomeView: UIView {
         
         homeWeeklyView.isHidden = false
         homeAllView.isHidden = true
+        
+        if !hasSlide {
+            homeWeeklyView.homeWeeklySlideYetView.isHidden = false
+            homeWeeklyView.homeWeeklySlidedView.isHidden = true
+        }
+        else {
+            homeWeeklyView.homeWeeklySlideYetView.isHidden = true
+            homeWeeklyView.homeWeeklySlidedView.isHidden = false
+        }
     }
     
     required init?(coder: NSCoder) {
@@ -43,12 +51,6 @@ final class HomeView: UIView {
     
     // MARK: - Custom Method
     
-    private func register() {
-        
-        weeklyCollectionView.register(thisWeekCell.self, forCellWithReuseIdentifier: thisWeekCell.cellIdentifier)
-        weeklyCollectionView.register(nextWeekCell.self, forCellWithReuseIdentifier: nextWeekCell.cellIdentifier)
-        
-    }
     
     private func style() {
         
@@ -58,42 +60,18 @@ final class HomeView: UIView {
         
         homeWeeklyView.do {
             $0.backgroundColor = .pppWhite
-            
-            if hasSlide {
-                $0.isHidden = true
-            }
-            else {
-                $0.isHidden = false
-            }
         }
         
         homeAllView.do {
             $0.backgroundColor = .pppWhite
         }
         
-        weeklyCollectionView.do {
-            let layout = AllCustomFlowLayout()
-            layout.scrollDirection = .horizontal
-            $0.collectionViewLayout = layout
-            $0.showsVerticalScrollIndicator = false
-            $0.isScrollEnabled = true
-            $0.translatesAutoresizingMaskIntoConstraints = false
-            $0.showsHorizontalScrollIndicator = true
-            $0.contentInsetAdjustmentBehavior = .never
-            
-            if hasSlide {
-                $0.isHidden = false
-            }
-            else {
-                $0.isHidden = true
-            }
-        }
         
     }
     
     private func hierarchy() {
         
-        self.addSubviews(homeWeeklyView, homeAllView, weeklyCollectionView, homeNavigationView)
+        self.addSubviews(homeWeeklyView, homeAllView, homeNavigationView)
         
     }
     
@@ -117,19 +95,14 @@ final class HomeView: UIView {
             $0.bottom.equalToSuperview().inset(Size.tabBarHeight)
         }
         
-        weeklyCollectionView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-            $0.width.equalTo(600)
-        }
-        
     }
     
     func dataBindArticleSlideCheck(articleData: HomeArticleCheckResult?) {
         print("😵‍💫실행됨")
         guard let hasSlide = articleData?.hasSlide else { return }
         self.hasSlide = hasSlide
-        self.homeWeeklyView.isHidden = hasSlide
-        self.weeklyCollectionView.isHidden = !hasSlide
+        self.homeWeeklyView.homeWeeklySlideYetView.isHidden = hasSlide
+        self.homeWeeklyView.homeWeeklySlidedView.isHidden = !hasSlide
     }
     
 }
