+++ PPPCLUB-iOS/Presentation/Home/ViewController/HomeViewController.swift
@@ -18,7 +18,7 @@ final class HomeViewController: BaseViewController{
     
     private var articleCardData: HomeArticleCardResult?  {
         didSet {
-            rootView.homeWeeklyView.dataBindArticleCard(articleData: articleCardData)
+            rootView.homeWeeklyView.homeWeeklySlideYetView.dataBindArticleCard(articleData: articleCardData)
         }
     }
     
@@ -65,7 +65,7 @@ final class HomeViewController: BaseViewController{
     private func gesture() {
         let gesture: UIPanGestureRecognizer = UIPanGestureRecognizer(target: self,
                                          action: #selector(ticketCaseMoved(_:)))
-        rootView.homeWeeklyView.ticketCoverImageView.addGestureRecognizer(gesture)
+        rootView.homeWeeklyView.homeWeeklySlideYetView.ticketCoverImageView.addGestureRecognizer(gesture)
     }
     
     private func target() {
@@ -76,8 +76,8 @@ final class HomeViewController: BaseViewController{
     private func delegate() {
         rootView.homeAllView.allArticleCollectionView.delegate = self
         rootView.homeAllView.allArticleCollectionView.dataSource = self
-        rootView.weeklyCollectionView.delegate = self
-        rootView.weeklyCollectionView.dataSource = self
+        rootView.homeWeeklyView.homeWeeklySlidedView.delegate = self
+        rootView.homeWeeklyView.homeWeeklySlidedView.dataSource = self
     }
     
     //MARK: - Action Method
@@ -103,7 +103,7 @@ final class HomeViewController: BaseViewController{
     }
     
     @objc private func ticketCaseMoved(_ sender: UIPanGestureRecognizer) {
-        AnimationManager.shared.ticketCoverAnimate(sender, targetView: rootView.homeWeeklyView.ticketCoverImageView) { _ in
+        AnimationManager.shared.ticketCoverAnimate(sender, targetView: rootView.homeWeeklyView.homeWeeklySlideYetView.ticketCoverImageView) { _ in
             self.ticketDragAnimation()
         }
     }
@@ -114,8 +114,8 @@ final class HomeViewController: BaseViewController{
 extension HomeViewController: UICollectionViewDelegateFlowLayout {
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
         switch collectionView {
-        case rootView.weeklyCollectionView:
-            let newtop = Size.height * 0.206
+        case rootView.homeWeeklyView.homeWeeklySlidedView:
+            let newtop = Size.height * 0.06
             let newbottom = Size.height * 0.212
             let newside = Size.width * 0.053 * 2
             return UIEdgeInsets(top: newtop, left: newside, bottom: newbottom, right: newside)
@@ -128,7 +128,7 @@ extension HomeViewController: UICollectionViewDelegateFlowLayout {
         switch collectionView {
         case rootView.homeAllView.allArticleCollectionView:
             return CGSize(width: 319, height: 180)
-        case rootView.weeklyCollectionView:
+        case rootView.homeWeeklyView.homeWeeklySlidedView:
             var cellHeight = Size.height * 0.58
             var cellwidth = Size.width * 0.786
             return CGSize(width: cellwidth, height: cellHeight)
@@ -141,7 +141,7 @@ extension HomeViewController: UICollectionViewDelegateFlowLayout {
         switch collectionView {
         case rootView.homeAllView.allArticleCollectionView:
             return 20
-        case rootView.weeklyCollectionView:
+        case rootView.homeWeeklyView.homeWeeklySlidedView:
             var spacingCal = Size.width * 0.026
             return spacingCal
         default:
@@ -157,7 +157,7 @@ extension HomeViewController: UICollectionViewDataSource {
         switch collectionView {
         case rootView.homeAllView.allArticleCollectionView:
             return articleAllData.count
-        case rootView.weeklyCollectionView:
+        case rootView.homeWeeklyView.homeWeeklySlidedView:
             return 2
         default:
             return 0
@@ -171,14 +171,14 @@ extension HomeViewController: UICollectionViewDataSource {
             cell.delegate = self
             cell.dataBindHome(articleData: articleAllData[indexPath.item])
             return cell
-        case rootView.weeklyCollectionView:
+        case rootView.homeWeeklyView.homeWeeklySlidedView:
             if indexPath.item == 0 {
-                guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: thisWeekCell.cellIdentifier, for: indexPath) as? thisWeekCell else { return UICollectionViewCell() }
+                guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: ThisWeekCell.cellIdentifier, for: indexPath) as? ThisWeekCell else { return UICollectionViewCell() }
                 cell.configureCell(articleData: articleCardData)
                 cell.delegate = self
                 return cell
             } else {
-                guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: nextWeekCell.cellIdentifier, for: indexPath) as? nextWeekCell else { return UICollectionViewCell() }
+                guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: NextWeekCell.cellIdentifier, for: indexPath) as? NextWeekCell else { return UICollectionViewCell() }
                 cell.configureCell(articleData: articleCardData)
                 return cell
             }
